Add new middleware registration to all meta packages.
