+++ src/SenseNet.Services.InMemory/SenseNetExtensions.cs
@@ -28,30 +28,33 @@ namespace SenseNet.Extensions.DependencyInjection
 
         /// <summary>
         /// Registers sensenet middlewares.
-        /// If you want to inject a middleware after the user was authenticated, use the onAfterAuthentication method parameter. 
+        /// If you want to inject a middleware before or after one of the built-in middlewares, use the
+        /// <see cref="MiddlewareBuilder"/> parameter for defining application builder methods. 
         /// </summary>
-        /// <remarks>
-        /// Please note that some of the middlewares (e.g. OData) are branching the pipeline. That means you cannot register
-        /// a custom middleware that runs after them. To be able to that, you have to call the Use methods in this method
-        /// directly and specify an app builder branch there.
-        /// </remarks>
         /// <param name="app">The application builder instance.</param>
-        /// <param name="onAfterAuthentication">An app builder method that can be used to register middlewares after
-        /// authentication but before the main sensenet middlewares (e.g. OData).</param>
-        public static IApplicationBuilder UseSenseNet(this IApplicationBuilder app, Action<IApplicationBuilder> onAfterAuthentication = null)
+        /// <param name="middlewareBuilder">Defines optional custom middlewares that will be registered before
+        /// or after certain sensenet middlewares (e.g. authentication or OData).</param>
+        public static IApplicationBuilder UseSenseNet(this IApplicationBuilder app, MiddlewareBuilder middlewareBuilder = null)
         {
-            // custom CORS policy
+            middlewareBuilder?.OnBeforeCors?.Invoke(app);
             app.UseSenseNetCors();
+            middlewareBuilder?.OnAfterCors?.Invoke(app);
 
-            // use Authentication and set User.Current
-            app.UseSenseNetAuthentication();
+            middlewareBuilder?.OnBeforeAuthentication?.Invoke(app);
+            app.UseSenseNetAuthentication(); // use Authentication and set User.Current
+            middlewareBuilder?.OnAfterAuthentication?.Invoke(app);
 
-            onAfterAuthentication?.Invoke(app);
-
+            middlewareBuilder?.OnBeforeMembershipExtenders?.Invoke(app);
             app.UseSenseNetMembershipExtenders();
-            app.UseSenseNetFiles();
-            app.UseSenseNetOdata();
-            app.UseSenseNetWopi();
+            middlewareBuilder?.OnAfterMembershipExtenders?.Invoke(app);
+
+            // conditional, terminating middlewares
+
+            app.UseSenseNetFiles(middlewareBuilder?.OnBeforeFiles, middlewareBuilder?.OnAfterFiles);
+
+            app.UseSenseNetOdata(middlewareBuilder?.OnBeforeOData, middlewareBuilder?.OnAfterOData);
+
+            app.UseSenseNetWopi(middlewareBuilder?.OnBeforeWopi, middlewareBuilder?.OnAfterWopi);
 
             return app;
         }
